python/requirements.txt
flask
flask-cors
mysql-connector-python
datetime